Refactor: export private chat mediator to a new file
